Server Scripts
Adding the current server scripts, untested, use at own risk.

These scripts cover the complete setup of a LEMP server and maintenance, as well as account management.

These set of bash script serve as a really good reference point for LEMP sys-admins.

A lot of the approaches in these scripts are old-school sysadmin, and are no longer used as in (in favour of DevOps style setups).